Split up config and services into separate files
Moved the different API services into their own files and the TMDB
config into its own file. This makes the code more modular and easier to
maintain.
