+++ src/index.ts
@@ -1,49 +1,13 @@
 import express, { Request, Response } from "express"
-import { generateHtmlTemplate } from "./templates/emailTemplate"
-
-const sgMail = require("@sendgrid/mail")
-const nodeMailer = require("nodemailer")
+import { TmdbResponse, tmdbData } from "./config/tmdbConfig"
+import { sendSendgridEmail } from "./services/sendgridService"
+import { sendSMTPEmail } from "./services/smtpService"
 
 const app = express()
 const port = process.env.PORT || 3000
 
 const baseUrl = "/webhook"
 const baseRouter = express.Router()
-
-interface TmdbBaseData {
-  apiKey: string
-  movieUrl: string
-  tvUrl: string
-  posterUrl: string
-  imdbUrl: string
-}
-
-interface TmdbResponse {
-  genres: string
-  homepage: string
-  id: number
-  overview: string
-  posterPath: string
-  releaseDate: string
-  runtime: number
-  tagline: string
-  title: string
-  movieUrl: string
-  imdbUrl: string
-}
-
-const tmdbData: TmdbBaseData = {
-  apiKey: process.env.TMDB_API_KEY || "",
-  movieUrl: "https://www.themoviedb.org/movie/",
-  tvUrl: "https://www.themoviedb.org/tv/",
-  posterUrl: "https://image.tmdb.org/t/p/original",
-  imdbUrl: "https://www.imdb.com/title/",
-}
-
-if (!tmdbData.apiKey) {
-  console.error("TMDB API Key not found")
-  process.exit(1)
-}
 
 app.use(express.json())
 
@@ -55,7 +19,7 @@ baseRouter.get("/", (req: Request, res: Response) => {
 
 baseRouter.post("/", async (req: Request, res: Response) => {
   printRequest(req)
-  if (!bodyContainsTmdbId(req.body)) {
+  if (!req.body.Provider_tmdb) {
     res.status(400).json({ message: "TMDB ID not found in body" })
     return
   }
@@ -85,16 +49,6 @@ function printRequest(req: Request) {
   console.log("Body: ", req.body)
 }
 
-function bodyContainsTmdbId(body: Request["body"]) {
-  // Provider_tmdb is the key that contains the TMDB ID
-  // This is sent when all details are sent to the webhook with no template and
-  // the jellyfin server uses tmdb for metadata.
-  if (!body.Provider_tmdb) {
-    return false
-  }
-  return true
-}
-
 async function fetchTmdbData(tmdbId: string) {
   const url: string = `https://api.themoviedb.org/3/movie/${tmdbId}?api_key=${tmdbData.apiKey}`
   const tmdbResponse = await fetch(url).then((response) => {
@@ -107,7 +61,6 @@ async function fetchTmdbData(tmdbId: string) {
     return response.json()
   })
 
-  console.log("TMDB Response: ", tmdbResponse)
   return tmdbResponse
 }
 
@@ -127,66 +80,6 @@ function formatTmdbResponse(response: any): TmdbResponse {
     movieUrl: `${tmdbData.movieUrl}${response.id}`,
     imdbUrl: `${tmdbData.imdbUrl}${response.imdb_id}`,
   }
-}
-
-async function sendSendgridEmail(formattedResponse: TmdbResponse) {
-  sgMail.setApiKey(process.env.SENDGRID_API_KEY)
-
-  const receiverEmails = (process.env.SENDGRID_RECEIVER_EMAIL || "")
-  .replace(/\s/g, "")
-  .split(",")
-  .map((email: string) => ({ email: email }))
-
-  const msg = {
-    from: {
-      email: process.env.SENDGRID_SENDER_EMAIL,
-    },
-    personalizations: [
-      {
-        to: receiverEmails, // array of objects with email key and value
-        dynamic_template_data: {
-          title: formattedResponse.title,
-          releaseDate: formattedResponse.releaseDate,
-          overview: formattedResponse.overview,
-          posterPath: formattedResponse.posterPath,
-          movieUrl: formattedResponse.movieUrl,
-          imdbUrl: formattedResponse.imdbUrl,
-        },
-      },
-    ],
-    templateId: process.env.SENDGRID_TEMPLATE_ID,
-  }
-  await sgMail.send(msg)
-}
-
-async function sendSMTPEmail(formattedResponse: TmdbResponse) {
-  const transporter = nodeMailer.createTransport({
-    host: process.env.SMTP_HOST,
-    port: process.env.SMTP_PORT || "587",
-    secure: false,
-    auth: {
-      user: process.env.SMTP_AUTH_USER,
-      pass: process.env.SMTP_AUTH_PASSWORD,
-    },
-  })
-
-  const mailOptions = {
-    from: process.env.SMTP_SENDER_EMAIL,
-    to: process.env.SMTP_RECEIVER_EMAIL,
-    subject:
-      "A new movie has been added to Jellyfin " +
-      formattedResponse.title +
-      ` (${formattedResponse.releaseDate})`,
-    html: generateHtmlTemplate(formattedResponse),
-  }
-
-  await transporter.sendMail(mailOptions, (error: any, info: any) => {
-    if (error) {
-      console.error(error)
-    } else {
-      console.log("Email sent: " + info.response)
-    }
-  })
 }
 
 app.listen(port, () => {
